requirements.txt
discord.py
aiomysql
aiohttp
python-dotenv
flask
requests
